refactor(#55): directory rendering & content-specific wrapper
